Temporarily remove broken wgpu 0.18 projects from workspace
wgpu 0.18 got yanked from crates.io so getting dependencies fails.
Fix requires upgrading to 0.19,
which is especially annoying for moonstaff
because that upgrade needs to be backported
to the version of starframe it's using.
Just removing them from the workspace for now
to be able to work on another project